Few fixups for Swallow Cmake. InteriorsExplorerViewModule - no longer takes a bunch of view models as the controller no longer needs them. (Controller was fixed up in previous iOS commit). Linking to sqlite extension. Fixup implicit casts. Buddy: Tim J

+++ src/Search/Swallow/SdkModel/SwallowSearchTransitionPinSelectionHandler.cpp
@@ -28,7 +28,7 @@ namespace ExampleApp
                 void SwallowSearchTransitionPinSelectionHandler::SelectPin()
                 {
                     Eegeo::Camera::CameraState cameraState = m_appCameraController.GetCameraState();
-                    float distanceToInterest = (cameraState.LocationEcef()-cameraState.InterestPointEcef()).Length();
+                    float distanceToInterest = static_cast<float>((cameraState.LocationEcef()-cameraState.InterestPointEcef()).Length());
                     m_transitionController.StartTransitionTo(m_transitionResult.GetTargetLatLong().ToECEF(), distanceToInterest,  m_transitionResult.GetTargetInteriorId(), m_transitionResult.GetTargetInteriorFloor());
                 }
             }
